+++ src/test/playground/ShouldExtendFullPaths.cs
@@ -1,0 +1,11 @@
+namespace TestAssets
+{
+    class ShouldExtendFullPaths
+    {
+        void x()
+        {
+            System.Collections.Generic.IEnumerable<string> x = null;
+            x.Equals("");
+        }
+    }
+}