+++ src/test/playground/ShouldNotShowExtensionsForStatic.cs
@@ -1,0 +1,10 @@
+using System.IO;
+namespace assets
+{
+    public class ShouldNotShowExtensionsForStatic
+    {
+        void x(){
+            File.WriteAllText("123","123");
+        }      
+    }
+}